New Shop UI
Started reconstructing all the UI in the game.

The new shop is a lot simpler and much easier to follow, understand and add too. Mostly due to much more use of USS instead of focusing on the visuals via C#.

All shop related things are now Scriptable objects and can very easily create themselves with a function call. The UI adapts better to screen adjustments and resolutions and also visually is more appealing.